着手写xml里的sql。下一步需要变更表状态，bak

+++ src/main/java/cn/ibaochenyu/jzh_shop/dao/entity/CommodityDO.java
@@ -23,5 +23,5 @@ public class CommodityDO {
     //1: 生产后入数据库记载，以商城有信息
     //2：锁定：用户下单。可能锁在支付付款
     //3：付款完成并给用户
-    private Long commodityStatus;
+    private int commodityStatus;
 }
